+++ user/api_v2/views.py
@@ -1,5 +1,7 @@
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import filters, status, viewsets
+from rest_framework.authtoken.models import Token
+from rest_framework.generics import get_object_or_404
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import BasePermission, IsAuthenticated
 from rest_framework.response import Response
@@ -8,6 +10,7 @@ from user.api_v2.serializers import (
     UserCreateSerializer,
     UserListSerializer,
     UserRetrieveSerializer,
+    UserTokenSerializer,
     UserUpdateSerializer,
 )
 from user.models import User
@@ -45,3 +48,19 @@ class UserAPI(viewsets.ModelViewSet):
         user.delete()
 
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class UserTokenAPI(viewsets.ModelViewSet):
+    serializer_class = UserTokenSerializer
+    permission_classes = [IsAuthenticated]
+
+    def retrieve(self, request):
+        instance = get_object_or_404(Token.objects.filter(user=request.user))
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
+
+    def refresh(self, request):
+        Token.objects.filter(user=request.user).delete()
+        instance = Token.objects.create(user=request.user)
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)